CORE-4216: Moved pipeline builder autobeans to de-commons
