+++ src/main/java/org/iplantc/de/client/views/windows/PipelineEditorWindow.java
@@ -1,12 +1,12 @@
 package org.iplantc.de.client.views.windows;
 
 import org.iplantc.de.client.models.WindowState;
+import org.iplantc.de.client.models.pipelines.Pipeline;
 import org.iplantc.de.client.views.windows.configs.ConfigFactory;
 import org.iplantc.de.client.views.windows.configs.PipelineEditorWindowConfig;
 import org.iplantc.de.client.views.windows.configs.WindowConfig;
 import org.iplantc.de.commons.client.info.IplantAnnouncer;
 import org.iplantc.de.commons.client.info.SuccessAnnouncementConfig;
-import org.iplantc.de.pipelineBuilder.client.json.autobeans.Pipeline;
 import org.iplantc.de.pipelines.client.presenter.PipelineViewPresenter;
 import org.iplantc.de.pipelines.client.views.PipelineView;
 import org.iplantc.de.pipelines.client.views.PipelineViewImpl;
